third_party/WebKit/LayoutTests/typedcssom/stylevalue-serialization/cssPositionValue.html

<!doctype html>
<meta charset="utf-8">
<title>IDL-constructed CSSPositionValue serialization tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#positionvalue-serialization">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
  assert_equals(new CSSPositionValue(CSS.px(1), CSS.percent(-3.14)).toString(), '1px -3.14%');
}, 'CSSPositionValue constructed from IDL serializes correctly');

test(() => {
  let result = CSSStyleValue.parse('object-position', '1px 1%');
  result.x = CSS.percent(-3.14);
  assert_equals(result.toString(), '-3.14% 1%');
  result.y = CSS.px(-3.14);
  assert_equals(result.toString(), '-3.14% -3.14px');
}, 'CSSPositionValue modified through "x" and "y" setters serializes correctly');

</script>
