third_party/WebKit/LayoutTests/typedcssom/stylevalue-serialization/cssPositionValue-expected.txt

This is a testharness.js-based test.
PASS CSSPositionValue constructed from IDL serializes correctly
FAIL CSSPositionValue modified through "x" and "y" setters serializes correctly assert_equals: expected "-3.14% 1%" but got "1px 1%"
Harness: the test ran to completion.
